Fixup elb helper
There was a change in TF to enforce empty instances in the elb if it was
set to be []. As the asg was adding things into the elb this led to them
being removed. The correct behaviour is to omit the key in the spec for
the elb.

Also refactored to allow spec to be any valid TF config, not plucking known
keys out